Found duplicated code and IFNE H6309-1 statements in dwwrite.asm.  Verified it was duplicated code.  Must have been accidently copied at some point.  Removed second copy of this IFNE H6309-1.
Changed H6309-1 to H6309 and moved the real H6309 code into this routine and moved the standard 6809 57600bps code to the last "else" section. This was done to dwread.asm, dwwrite.asm, dw4write.asm, and dw4read.asm.

Test compiled and everything still seems to compile as it should.