Added becker archive creation to the main makefile.  The command is "make becker".
Also made a change in the level2/coco3/makefile_common to change the file creation related to the becker disks not to be referred to as 80 track double sided disk images as this is not correct.  The DriveWire4 server requires the disk images to be single sided for the HDBDOS "DOS" command to work.  Corrected makefile so the 80d is not part of the filename.