Found ENDC in wrong location in dwread.asm and dwwrite.asm.  Corrected.
Moved the native 6309 code in dwread.asm and dwwrite.asm into the H6309 labeled area and changed IFEQ H6309 to IFNE H6309.  Also moved the 57600bps 6809 code to the default location.  This change had been done in the old  dwread.asm and dwwrite.asm files to make it easier to follow.  Though these two files were overwritten from the HDBDOS project dwread.asm and dwwrite.asm files.  So this conversion needed to be done again so it made the source easier to follow.