Makefiles: Do not let $(RM) silently fail
The - in front of $(RM) was probably remains from
a time when RM was not rm -f, and is not needed any
longer.

In general, if something fails we should notice and fix it, so
using the preceding "-" in makefiles is not a good practice.
The only occurences left are the SOFTLINKs in a few sys/makefile's.