rb1773: Make 48 TPI floppy disks readonly in 96 TPI drives
Deny the disk destroying write with the narrower heads of the 96 tpi drive
when the user has placed a 48 tpi disk in the 96 tpi drive so he can read
it. Double stepping the drive is then done automatically so as to stay in
mechanical synch with the tracks on a 48 tpi disk. But that, until now, did
not make the disk read-only, this patch does that.

Patch by Gene Heskett and Robert Gault.