mnln.asm: Fix obj.status.v command calling format text
I also fixed the obj.status.v command (L0BF4). It's a debugging
command used to display information about animated objects shown on the
screen. The current implementation displays garbled text. The command
calls a routine to format a string with arguments (L3ED6). The arguments
to format are pushed on the stack before calling the formatting routine
but two of them were incorrectly assigned.