level2/6309 bootfile script: Fix typo in options
Thanks to Robert Gault for reporting. The typo prevents the script
from running and is almost impossible to see when you "list mb.dw"
as the first line scrolls off the screen before you see it.

Also modify the filename in a message to make it consistent with
other scripts. This script should be identical to the 6809 one.