Fix idiocy related to pshu/pulu with "u" as an operand
Historically, lwasm and some other assemblers silently accept "u" as a
register to pshu/pulu. The problem is that it is impossible to encode in
the object code since the bit pattern for U and S are identical and is
treated as U for PSHS/PULS and S for PSHU/PULU.

Replace all instances of "U" in PSHU/PULU with "S". This is necessary due to
upcoming tightening of code validation in lwasm, and also makes it properly
clear what is really happening for people who aren't intimately familiar
with the intricacies of the 6809.