makefiles: Use port.mak for level 2 and 3 as well
The port.mak defines the OS level, CPU arch etc so
those are not defined in the makefiles. This allows
sharing of makefiles between different ports.

For now the level 3 build is limited to a few disk images,
and has its own makefile. At a later point we can hopefully
share makefiles between level 2 and 3, checking LEVEL when needed.