Use lwtools noforwardrefmax pragma by default
The in-progress development version of lwasm changes the default value
of forwardrefmax with the effect that some pc-relative forward branches
end up as 16-bit when they could be 8-bit. This change in behaviour can
cause mismatched binaries between builders using version 4.13 and
builders using later versions. This can have a particularly bad effect
on builds of level2 kernels, because the current code has hand-inserted
padding to align the end-point of the code. The fix for all of this is
to change the command-line options to force the old default behaviour.
http://lwtools.projects.l-w.ca/hg/index.cgi/rev/b4825b42c151