defs/cocovtio.d: Add ModCo80 COLoad bit mask for Co80 VDIO submodule
This is for the CRT9128-based Co80 driver, not for the 6845-based CoWP
driver (WordPak I/II/RS).

The Co80 driver name has at some point also been used for the
WordPak RS driver, however co80.asm is for a different chip, the
CRT9128, used (at least) in a PBJ prototype.

We later got a cowprs.asm driver for the WorkPak RS. The module name
is CoWP and it uses the ModCoWP bit mask.

Since VTIO (level 1) has hardcoded module names for the submodules,
matched to a COLoad bit mask value, these two modules must have
different COLoad bit masks to allow both to exist on the same system.