Get gfx2 to build from cmds folders
Now that gfx2.asm needs to include coco3vtio.d the latter needs to be
built first and be accessed from the port-specific defs folder.

OTOH the one built in 3rd-party/packages/basic09 is built only later
during a full build.

(All the different defs folders and .d files still puzzle me.)

Also updated coco3fpga/defs/makefile to be based on the coco3 one.

The mc09l2 port doesn't need gfx and gfx2.

Note for later:
This whole build of basic09 components from the 3rdparty folder
during the ports builds is not very nice. At a minimum we should
try to limit the extra --includedir to the few commands needing it.